Semi-support for multiple game images. Each <image> in gamelist.xml now supports an "id" attribute, e.g.
	<image id="0">/home/framps/wucade/mame/snaps/dkong.png</image>
	<image id="1">/home/framps/wucade/mame/marquees/dkong.png</image>

Images without an "id" are deprecated. If an image is found without an id it will be renamed
to id="0". This will cause an issue if someone attempts to make a non-id image, and an image
with id="0" in the same game. This shouldn't occur if all scraping is done with ids in the
future.

At the moment the actual images are not being created, need to add suppot for multiple
game images in the theme.